refactor: simplify test structure, ensure atomicity of test workflows, exploit synergies
