+++ packages/graphql/test/userData.ts
@@ -1,0 +1,39 @@
+// mock user 1
+export const userOne = {
+  id: 'b086deed-291d-4dc3-9271-c84d4c2840f2',
+  sub: 'b086deed-291d-4dc3-9271-c84d4c2840f2',
+  email: 'first@example.com',
+  shortname: 'firstuser',
+}
+
+// mock user 2
+export const userTwo = {
+  id: '5deec433-7ead-4c9c-9f6d-96cd5ef1758a',
+  sub: '5deec433-7ead-4c9c-9f6d-96cd5ef1758a',
+  email: 'second@example.com',
+  shortname: 'seconduser',
+}
+
+// mock user 3
+export const userThree = {
+  id: 'e9a2e2f8-6d1d-4a6d-8d7f-7f3d4d4a1f4a',
+  sub: 'e9a2e2f8-6d1d-4a6d-8d7f-7f3d4d4a1f4a',
+  email: 'third@example.com',
+  shortname: 'thirduser',
+}
+
+// mock user 4
+export const userFour = {
+  id: 'f3c7d4f3-4e3d-4f3d-4f3d-4f3d4f3d4f3d',
+  sub: 'f3c7d4f3-4e3d-4f3d-4f3d-4f3d4f3d4f3d',
+  email: 'fourth@example.com',
+  shortname: 'fourthuser',
+}
+
+// mock user 5
+export const userFive = {
+  id: 'f3c7d4f3-4e3d-4f3d-4f3d-4f3d4f3d4f3e',
+  sub: 'f3c7d4f3-4e3d-4f3d-4f3d-4f3d4f3d4f3e',
+  email: 'fifth@example.com',
+  shortname: 'fifthuser',
+}
